Make picante pass R CMD check
Current R CMD check performs more stringent tests on the import and
use of functions in external packages. A lightway method of passing
R CMD check is to explicitly refer to the NAMESPACE of these
packages, e.g., quantreg::rq(). A heavier alternative (in another
branch) is to change dependencies in DESCRIPTION and import these
functions, so that NAMESPACE would have lines like

    importFrom(quantreg, rq)

and to remove require(quantreg) in the function body.